Move the cloud uplink out of pkg/anywhere
pkg/anywhere held two different things. The control plane — one
persistent payload-agnostic link per cluster, which RFD-94 calls the
unified uplink and treats as the backbone for all runtime↔cloud
coordination. And the data plane — POPManager and connection.request,
many connections, existing to forward public traffic into a NAT'd
cluster. They couple at exactly one seam, where the data plane's setup
handshake is delivered over the control plane.

Left alone, every future tenant would import a package named for NAT
traversal to reach cloud, and deploy reporting was about to be the
second. Renaming the package wouldn't have helped, since POPManager
genuinely is Anywhere.

So the control plane moves to pkg/uplink and Anywhere becomes a tenant
of it, holding a client it no longer owns. coordinate owns the link,
because several features share it and none of them should be reached
through another. Time sync and org info move too: both are properties
of the link rather than of Anywhere, and cloud reconciles reported
timestamps against the offset, so a reporter needing a clock should not
have to ask the NAT-traversal package for one.

The app.snapshot payload moves the other way, out of the pipe and in
with the reporter that owns it. That direction matters as much as the
other: the link should keep understanding nothing about apps, and the
types sitting in its package were the beginning of it learning.

No behavior change. Verified against a local cloud end to end — the
cluster connects, syncs its clock, resolves its org, and reports app
health, now logging under component=uplink rather than anywhere.

Closes MIR-1568
